✨Code cleanup (#47)
- Sorts import order
- Uses short-syntax for fragments (`<>`)
- Remove unused components (`FilterCategory.js`)
- Moves App.js out of `/components` into `src`
- Change redirect flow for Login and Resources to use `Redirect`
- Rename components
  - `Filter.js` to `Resources.js`
  - `RegisterForm.js` to `Register.js`
  - `ResourceBreadcrumb.js` to `ResourcesBreadcrumb.js`
  - `ResourceViewFilterHeader.js ` to `ResourcesFilter.js`
  - `ResourceViewGrid.js ` to `ResourcesGrid.js`
  - `ResourceViewHeader.js` to `ResourcesBanner.js`
