+++ client/src/components/AdminResourceManager.js
@@ -1,6 +1,7 @@
-import React, { useState } from 'react';
-import '../css/AdminResourceManager.css';
+import React from 'react';
+
 import NewResourceForm from './NewResourceForm';
+import '../css/AdminResourceManager.css';
 
 const AdminResourceManager = () => {
   return (
