+++ client/src/components/Home.js
@@ -1,8 +1,9 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
-import { Row, Col, Button, Carousel } from 'antd';
-import '../css/Home.css';
 import { Textfit } from 'react-textfit';
+import { Button, Carousel, Col, Row } from 'antd';
+import '../css/Home.css';
+
 import headerImg from '../imgs/homePage/headerImg.jpg';
 
 const gradientBox = {
@@ -66,7 +67,7 @@ export default class Home extends Component<Props, State> {
 
   render() {
     return (
-      <div>
+      <>
         <Row
           className="home-block-1"
           type="flex"
@@ -227,7 +228,7 @@ export default class Home extends Component<Props, State> {
             </Row>
           </Col>
         </Row>
-      </div>
+      </>
     );
   }
 }
