+++ client/src/components/NewResourceAvailableLangForm.js
@@ -1,7 +1,7 @@
 // @flow
 
 import React, { useState } from 'react';
-import { Form, Input, Button, List } from 'antd';
+import { Button, Form, Input, List } from 'antd';
 
 type FormProps = {
   availableLanguages: () => void,
