--- client/src/css/ResourceBreadcrumb.css
@@ -1,9 +1,0 @@
-.link {
-    color: white;
-}
-
-.link:hover {
-    color: white;
-    font-weight: bold;
-    text-decoration: none;
-}
