--- client/src/css/ResourceViewFilterHeader.css
@@ -1,49 +1,0 @@
-.button {
-    border-radius: 20em;
-    width: 100px;
-}
-
-.button:hover {
-    background-color: #431C72;
-    border-color: white;
-    color: white;
-}
-
-.dropdown {
-    margin-left: 2em;
-}
-
-.header-container {
-    background: white;
-    border-top: 1px solid black;
-    padding-left: 220px;
-    padding-top: 3vh;
-    padding-bottom: 3vh;
-}
-
-.radio {
-    display: block;
-    height: 3em;
-    line-height: 3em;
-}
-
-.radio-container {
-    background: white;
-    box-shadow: 1px 1px 3px grey;
-    padding-bottom: 1em;
-    padding-left: 3em;
-    padding-right: 3em;
-    padding-top: 1em;
-    text-align: left;
-}
-
-.searchbar {
-    margin-right: 2em;
-    float: right;
-    padding-right: 1.5em;
-    width: 300px;
-}
-
-.title {
-    text-align: center;
-}
